feat: validate GitHub Actions workflows as part of build (#18)
The broken workflow as a result of the typo introduced in #15 surprised
me that this is one area where Projen doesn't "save you from yourself",
so to speak. This brings in https://github.com/mpalmer/action-validator
to validate that any workflows generated by Projen are indeed valid. I
chose to make it a post-compile task so that this check fails when
you're running a build locally, and not just in CI (although it will
also run in CI).