chore!: upgrade to cdktf 0.21.0 (#123)
This PR initiates the upgrade of CDKTF from version `0.20.0` to version
`0.21.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-null` to a version compatible with
`cdktf@0.21.0`
[here](https://github.com/cdktf/construct-projen-template/blob/d62067602139725e957e516ef41d94c384af731a/.projenrc.ts#L84).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-null/releases/).
- [x] Run `npx projen`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>